# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación para la API de tareas (INGCO)s",
        "contact": {
            "email": "meifer.elitepvpers@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tasks": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Obtener todas las tareas",
                "description": "Retorna todas las tareas disponibles",
                "operationId": "getTasksAPI",
                "responses": {
                    "200": {
                        "description": "Lista de tareas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tasks": {},
                                        "message": {
                                            "type": "string",
                                            "example": "Tareas obtenidas correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/tags": {
            "get": {
                "tags": [
                    "Tags"
                ],
                "summary": "Obtener todas las etiquetas",
                "description": "Retorna todas las etiquetas disponibles",
                "operationId": "getTagsAPI",
                "responses": {
                    "200": {
                        "description": "Lista de etiquetas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "tags": {},
                                        "message": {
                                            "type": "string",
                                            "example": "Etiquetas obtenidas correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/task": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Crear una nueva tarea",
                "description": "Crea una nueva tarea con la información proporcionada",
                "operationId": "createTaskAPI",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "due_date",
                                    "user_id",
                                    "tags"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nombre de la tarea"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción de la tarea"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01 12:00:00"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tags": {
                                        "type": "string",
                                        "example": "1,2,3,6"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "task": {},
                                        "message": {
                                            "type": "string",
                                            "example": "Tarea creada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "El campo name es requerido."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Obtener todos los usuarios",
                "description": "Retorna todos los usuarios disponibles",
                "operationId": "getUsersAPI",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "users": {},
                                        "message": {
                                            "type": "string",
                                            "example": "Usuarios obtenidos correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/api/task/{task}": {
            "put": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Actualizar una tarea existente",
                "description": "Actualiza una tarea existente con la información proporcionada",
                "operationId": "updateTaskApi",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID de la tarea a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "due_date",
                                    "user_id",
                                    "tags"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Nombre actualizado de la tarea"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descripción actualizada de la tarea"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01 12:00:00"
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tags": {
                                        "type": "string",
                                        "example": "1,2,3,6"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarea actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "task": {},
                                        "message": {
                                            "type": "string",
                                            "example": "Tarea actualizada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "El campo name es requerido."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/tasks/{task}/delete": {
            "delete": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Eliminar una tarea",
                "description": "Elimina una tarea existente",
                "operationId": "deleteTaskAPI",
                "parameters": [
                    {
                        "name": "task",
                        "in": "path",
                        "description": "ID de la tarea a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarea eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Tarea eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        }
    },
...
}